Move download handler function into handlers
and rename it to follow *_handler naming of the other two
